Remove content negotiation
Provide the plain text listing and JSON structure via separate resources
that always returns the same content regardless of who requests it.  The
use of content negotiation made it impossible to easily debug the
outputs of the other formats by loading it in the browser as the browser
would never receive them.