Implement downloading snapshots
Provide an easy interface for downloading a packed file will one save
from each of the instances that happened at a certain point in time.

This makes playing the cluster much easier, as you don't have to script
fetching the files, or download all 500 GB to get them in a format that
can easily loaded in either Factorio or Clusterio.